requirements.txt
# requirements.txt
google-generativeai
python-dotenv
PyMuPDF
Pillow